Disable practice matches for 2015
The 2015 data will come from The Blue Alliance, which only reports
qualification and elimination matches AFAIK.  The best workaround
I think is to just disable the option to view practice matches
when 2015 is selected.

This shouldn't be too much of a problem, since practice scores
rarely provide any insight.  It's just kind of funny to see fake
scores in the millions.